backend/requirements.txt
fastapi==0.127.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.45
asyncpg==0.31.0
alembic==1.17.2
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
pyyaml==6.0.3
types-PyYAML==6.0.12.20250915
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
httpx==0.28.1
aiosqlite==0.22.1
mypy==1.19.1
